requirements.txt
firebase_admin
Flask
crypto
cryptography
Fernet
discord-py-interactions